fix merge

+++ sources/Plugins/Primitives/Cylinders/CylinderEntryPoint.cpp
@@ -23,11 +23,11 @@ extern "C" {
         if (setting.exists("position")){
             const RayTracer::Shared::SettingWrapper &settingA = setting.lookup<RayTracer::Shared::SettingWrapper>("position");
             if (settingA.exists("x") && settingA.exists("y") && settingA.exists("z")) {
-                int x, y, z;
+                float x, y, z;
                 try {
-                    x = static_cast<int>(settingA.lookup<int>("x"));
-                    y = static_cast<int>(settingA.lookup<int>("y"));
-                    z = static_cast<int>(settingA.lookup<int>("z"));
+                    x = static_cast<float>(settingA.lookup<float>("x"));
+                    y = static_cast<float>(settingA.lookup<float>("y"));
+                    z = static_cast<float>(settingA.lookup<float>("z"));
                 } catch (const libconfig::SettingTypeException& ex) {
                     throw RayTracer::Shared::ConfigError("Cylinder", "Unable To read position coordinates");
                 }
@@ -48,11 +48,11 @@ extern "C" {
         if (setting.exists("rotation")) {
            const RayTracer::Shared::SettingWrapper &settingB = setting.lookup<RayTracer::Shared::SettingWrapper>("rotation");
             if (settingB.exists("x") && settingB.exists("y") && settingB.exists("z")) {
-                int x, y, z;
+                float x, y, z;
                 try {
-                    x = static_cast<int>(settingB.lookup<int>("x"));
-                    y = static_cast<int>(settingB.lookup<int>("y"));
-                    z = static_cast<int>(settingB.lookup<int>("z"));
+                    x = static_cast<float>(settingB.lookup<float>("x"));
+                    y = static_cast<float>(settingB.lookup<float>("y"));
+                    z = static_cast<float>(settingB.lookup<float>("z"));
                 } catch (const libconfig::SettingTypeException& ex) {
                     throw RayTracer::Shared::ConfigError("Cylinder", "Unable To read rotation coordinates");
                 }
